+++ client/src/data.js
@@ -1,0 +1,12 @@
+export default {
+  user: {
+    handle: 'mg',
+    email: 'mohammedgaber273@gmail.com',
+    imageUrl:
+      'https://firebasestorage.googleapis.com/v0/b/social-app0.appspot.com/o/174767524904929.png?alt=media',
+    location: 'Cairo, Egypt',
+    bio: 'front end developer',
+    website: 'http://mg.com',
+    userId: 'TZLqrWg4aoTHVbacWg5ofcwIaFW2',
+  },
+}