+++ client/src/redux/action-types.js
@@ -1,0 +1,12 @@
+export const REGISTER_FAIL = 'REGISTER_FAIL'
+export const REGISTER_SUCCESS = 'REGISTER_SUCCESS'
+
+export const LOGIN_FAIL = 'LOGIN_FAIL'
+export const LOGIN_SUCCESS = 'LOGIN_SUCCESS'
+
+export const AUTH_ERROR = 'AUTH_ERROR'
+export const LOAD_USER = 'LOAD_USER'
+export const USER_LOADED = 'USER_LOADED'
+export const LOAD_USER_FAILURE = 'LOAD_USER_FAILURE'
+
+export const USER_LOGOUT = 'USER_LOGOUT'