+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 #### CLIENT
 
-- added Navbar
-- added HomePage
-- added Post 
+- Added LoginPage and SignupPage
+- Redux: setup store and auth reducer (with user info)
+- Navbar: check if user logged-in and show user details and logout
 
 #### SERVER